[fix]: rename some files
