+++ web/src/config/chartConfig.ts
@@ -1,5 +1,5 @@
 import { UnitEnum } from '../model/Metric'
-import { DataFormatter } from '../utils/dataFormatter'
+import { DataFormatter } from '../utils/DataFormatter'
 
 const ChartJS = require('chart.js')
