+++ web/src/containers/Setting/Database.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { autobind } from 'core-decorators'
 import { Button, Card, Table } from 'antd'
 import { redirectTo } from '../../utils/URLUtil'
-import { DatabaseInfo, getDatabaseList, } from '../../service/storage'
+import { DatabaseInfo, getDatabaseList, } from '../../service/Storage'
 import { observable } from 'mobx'
 import { observer } from 'mobx-react'
 import ReactJson from 'react-json-view'
