+++ web/src/service/Login.ts
@@ -1,4 +1,4 @@
-import { POST } from './http'
+import { POST } from './APIUtils'
 import { PATH } from '../config/config'
 
 export function login(username: string, password: string) {
